# fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GymBuddy API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "parameters": [
          {
            "name": "demo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/occupancy": {
      "get": {
        "summary": "Occupancy",
        "operationId": "occupancy_occupancy_get",
        "parameters": [
          {
            "name": "demo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccupancyResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/forecast": {
      "get": {
        "summary": "Forecast",
        "operationId": "forecast_forecast_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FacilityId"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Id"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          },
          {
            "name": "demo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/recommend": {
      "post": {
        "summary": "Recommendation",
        "operationId": "recommendation_recommend_post",
        "parameters": [
          {
            "name": "demo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Demo"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Candidate": {
        "properties": {
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "facility_id": {
            "$ref": "#/components/schemas/FacilityId"
          },
          "facility_name": {
            "type": "string",
            "title": "Facility Name"
          },
          "predicted_occupancy_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Predicted Occupancy Pct"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence"
          },
          "provenance": {
            "type": "string",
            "enum": [
              "demo",
              "live",
              "cached",
              "unavailable"
            ],
            "title": "Provenance"
          },
          "exceeds_tolerance": {
            "type": "boolean",
            "title": "Exceeds Tolerance",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_time",
          "end_time",
          "facility_id",
          "facility_name",
          "predicted_occupancy_pct",
          "confidence",
          "provenance"
        ],
        "title": "Candidate"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "FacilityForecast": {
        "properties": {
          "forecast_source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "databricks",
                  "collector_fallback"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Source"
          },
          "facility_id": {
            "$ref": "#/components/schemas/FacilityId"
          },
          "facility_name": {
            "type": "string",
            "title": "Facility Name"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ForecastPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence"
          },
          "provenance": {
            "type": "string",
            "enum": [
              "demo",
              "live",
              "cached",
              "unavailable"
            ],
            "title": "Provenance"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "facility_id",
          "facility_name",
          "points",
          "generated_at",
          "observed_at",
          "confidence",
          "provenance"
        ],
        "title": "FacilityForecast"
      },
      "FacilityId": {
        "type": "string",
        "enum": [
          "mccomas",
          "war_memorial"
        ],
        "title": "FacilityId"
      },
      "ForecastPoint": {
        "properties": {
          "forecast_time": {
            "type": "string",
            "format": "date-time",
            "title": "Forecast Time"
          },
          "predicted_occupancy_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Predicted Occupancy Pct"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "forecast_time",
          "predicted_occupancy_pct"
        ],
        "title": "ForecastPoint"
      },
      "ForecastResponse": {
        "properties": {
          "facilities": {
            "items": {
              "$ref": "#/components/schemas/FacilityForecast"
            },
            "type": "array",
            "title": "Facilities"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Data Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "facilities",
          "data_mode"
        ],
        "title": "ForecastResponse"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded"
            ],
            "title": "Status"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Data Mode"
          },
          "integrations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/IntegrationStatus"
            },
            "type": "object",
            "title": "Integrations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "data_mode",
          "integrations"
        ],
        "title": "HealthResponse"
      },
      "IntegrationStatus": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "status": {
            "type": "string",
            "enum": [
              "not_configured",
              "not_implemented",
              "ready",
              "unavailable"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "configured",
          "status"
        ],
        "title": "IntegrationStatus"
      },
      "Interval": {
        "properties": {
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_time",
          "end_time"
        ],
        "title": "Interval"
      },
      "Occupancy": {
        "properties": {
          "facility_id": {
            "$ref": "#/components/schemas/FacilityId"
          },
          "facility_name": {
            "type": "string",
            "title": "Facility Name"
          },
          "occupancy": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupancy"
          },
          "remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "occupancy_pct": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupancy Pct"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "source_updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Updated At"
          },
          "provenance": {
            "type": "string",
            "enum": [
              "demo",
              "live",
              "cached",
              "unavailable"
            ],
            "title": "Provenance"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "facility_id",
          "facility_name",
          "provenance"
        ],
        "title": "Occupancy"
      },
      "OccupancyResponse": {
        "properties": {
          "facilities": {
            "items": {
              "$ref": "#/components/schemas/Occupancy"
            },
            "type": "array",
            "title": "Facilities"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Data Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "facilities",
          "data_mode"
        ],
        "title": "OccupancyResponse"
      },
      "RecommendationRequest": {
        "properties": {
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "unavailable": {
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Unavailable"
          },
          "workout_duration_minutes": {
            "type": "integer",
            "maximum": 180.0,
            "minimum": 15.0,
            "title": "Workout Duration Minutes",
            "default": 75
          },
          "preferred_gyms": {
            "items": {
              "$ref": "#/components/schemas/FacilityId"
            },
            "type": "array",
            "maxItems": 2,
            "title": "Preferred Gyms"
          },
          "crowd_tolerance": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Crowd Tolerance",
            "default": "low"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_time",
          "end_time"
        ],
        "title": "RecommendationRequest"
      },
      "RecommendationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "no_available_window",
              "data_unavailable"
            ],
            "title": "Status"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Candidate"
              },
              {
                "type": "null"
              }
            ]
          },
          "alternative": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Candidate"
              },
              {
                "type": "null"
              }
            ]
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "method": {
            "type": "string",
            "enum": [
              "deterministic",
              "gemini"
            ],
            "title": "Method",
            "default": "deterministic"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Data Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "explanation",
          "generated_at",
          "data_mode"
        ],
        "title": "RecommendationResponse"
      }
    }
  }
}